==> 5-Geocoding/0-testing.ipynb <==
import pandas as pd
#ksdm
#lsakdckmd

df_q = pd.read_csv('data/oda-addresses/ODA_QC_v1.csv', low_memory=False)

# what's the problem with quebec matching?
test = df_q[df_q.street == 'Rue Queen']

# how many results are there to geocode?

df = pd.read_csv('formatted.csv', low_memory=False)

len(df)

list(df)

df = df[df['latitude'].isnull()]

len(df)

df = pd.read_csv('data/4-ODBiz_merged_2022-08-04.csv')

len(df)

df = df[df['longitude'].isnull()]

len(df)

# number without lat/ lon: 598891

# I might want to deduplicate before running the address match script
df = df.drop_duplicates(subset=['street_no','formatted_en'])

df = pd.read_csv('output_all.csv', low_memory=False)

# get those which match lat lon

df_lat = df[~df['latitude'].isna()]
df_lat = df_lat[~df_lat['x'].isna()]

df_lat.to_csv("latlon_compairson.csv", index=False)

# get sample of imperfect matches between 60-95

# df_match = df_q[df_q.ratio == 'Rue Queen']
df_match = df[df['ratio'].between(60, 98)]


# in general, 80 good, but not for numbers (eg 10th st versus 15th st are very similar)

df_match

df_match.to_csv("imprefect_match_sample.csv", index=False)

df = pd.read_csv('data/ODBiz_parsed.csv')
df['full_address'].isna().sum()


len(df)

# results
