==> 5-Geocoding/2-oda-download.ipynb <==
# download the ODA data files

import requests, zipfile, io

# go 

provinces = ['AB', 'BC', 'MB', 'NB', 'NT', 'NS', 'ON', 'PE', 'QC', 'SK']

for province_code in provinces:
    zip_file_url = "https://www150.statcan.gc.ca/n1/pub/46-26-0001/2021001/ODA_" + province_code + "_v1.zip"
    r = requests.get(zip_file_url)
    z = zipfile.ZipFile(io.BytesIO(r.content))
    z.extractall("data/oda-addresses")
    



# https://www150.statcan.gc.ca/n1/pub/46-26-0001/2021001/ODA_AB_v1.zip

# https://www150.statcan.gc.ca/n1/pub/46-26-0001/2021001/ODA_BC_v1.zip
    
# https://www150.statcan.gc.ca/n1/pub/46-26-0001/2021001/ODA_SK_v1.zip

import pandas as pd
from Address_Format_Funcs import AddressClean_en, AddressClean_fr

# apply formatting to quebec\
# CHANGED apply formatting script to all provinces


provinces = ['AB', 'BC', 'MB', 'NB', 'NT', 'NS', 'ON', 'PE', 'QC', 'SK']


for province_code in provinces:
    print(province_code)
    ocd_file = "data/oda-addresses/ODA_" + province_code + "_v1.csv"
    df = pd.read_csv(ocd_file, low_memory=False)
    if (province_code == 'QC'):
        df = AddressClean_fr(df,'street','street')
    else:
        df = AddressClean_en(df,'street','street')
    output_file_name = "data/oda-addresses/ODA_" + province_code + "_v1_formatted.csv"
    df.to_csv(output_file_name, index=False)

# list(df)
# df = df.sample(30)
df['street'].tail(30)

df = AddressClean_fr(df,'street','street')
# df['street']=df['formatted_fr']
df['street']=df['street'].str.lower()

df.street = df.street.str.replace('  ', ' ')

df.to_csv('data/oda-addresses/ODA_QC_v1_formatted.csv',index=False)
