==> 5-Geocoding/3-deduplication.ipynb <==
# We're doing some basic deduplication to remove exact matches for street number 
# + strong fuzzy matches for street names
# within the same city


# NOTE if I'm deduplicating, I need to be careful to retain data
# specifically for now, I want to keep the one with a geocode

# I'll need to sort the data first perhaps - https://stackoverflow.com/questions/43814487/remove-duplicate-rows-in-pandas-dataframe-based-on-condition
# then the nan ones for geo will be lower

import pandas as pd

from thefuzz import fuzz
from thefuzz import process

df = pd.read_csv('data/formatted.csv', low_memory=False)
print('rows: ', len(df))

print(len(df[df['longitude'].isnull()]))

df['latitude'].tail(30)

# Start by removing exact matches

# df = df.drop_duplicates(subset=['street_no','formatted_en', 'province'])
# df = df[(~df.duplicated(subset=['street_no','formatted_en', 'province'])) | df['street_no','formatted_en', 'province'].isna()]

# sort so that we keep records of the lat/lon if present
df.sort_values(['latitude'])
# this only seems to change the result by 20. Does that make sense? That suggests that pretty much all of the duplicates have the same lat/lon information

df_city = df[~df.duplicated(subset=['street_no','formatted_en', 'province', 'city'], keep='first') | df['street_no'].isna() | df['formatted_en'].isna() | df['province'].isna() | df['city'].isna()]
# df_province = df[~df.duplicated(subset=['street_no','formatted_en', 'province']) | df['street_no'].isna() | df['formatted_en'].isna() | df['province'].isna()]
# df_nogeo = df[~df.duplicated(subset=['street_no','formatted_en']) | df['street_no'].isna() | df['formatted_en'].isna()]

# FOR TESTING create a dataframe of duplicates for inspection

print('rows: ', len(df_city))
# print('rows: ', len(df_province))
# print('rows: ', len(df_nogeo))
# if (province_code == 'QC'):
#     df = df.drop_duplicates(subset=['street_no','formatted_fr'])
# else:
#     df = df.drop_duplicates(subset=['street_no','formatted_en'])

duplicates = df[df.duplicated(subset=['street_no','formatted_en', 'province', 'city'], keep=False)]

duplicates = duplicates.sort_values(['street_no','formatted_en', 'province', 'city'])
duplicates[['street_no','formatted_en', 'province', 'city']].head(30)

# how many to geocode?
print(len(df_city[df_city['latitude'].isnull()]))

df_city['province'].unique()

df_city[(df_city['province'] == 'LAKEVIEW ESTATES') | (df_city['province'] == '78')]

len(df)

df['full_address'].isna().sum()

list(df)
