==> 5-Geocoding/5-results.ipynb <==
# Results analysis

# Took X hours
# Overall X 100% matches and X matches above 80%
# Yet to implement matches based on city - but so far looks good. X % fuzzy matches
# Around X left to geocode, depending on our cutoff point
# X for 80% cutoff
# X for 85% cutoff
# X for 90% cutoff

import pandas as pd

df = pd.read_csv('output_all.csv', low_memory=False)

len(df)

# get number of matches above various cutoffs for the entire dataset
cut_off = 99.5

for number in [80,85,90,95,99.5]:
    no_matches = df['ratio'][df['ratio'] > number].count()
    percent_matches = 100 * no_matches / len(df)
    print('Matches above ', number, '%: ', int(percent_matches), '%', sep="")

# no_matches_80 = df['ratio'][df['ratio'] > 80].count()
# no_matches_85 = df['ratio'][df['ratio'] > 85].count()
# no_matches_90 = df['ratio'][df['ratio'] > 90].count()
# no_matches_95 = df['ratio'][df['ratio'] > 95].count()
# no_matches_100 = df['ratio'][df['ratio'] > 99.5].count()

# print('matches above' no_matches_80)


# number left to geocode

df_lat = df[df['latitude'].isna()]
print('original number to geocode:', len(df_lat))
df_lat = df_lat[df_lat['x'].isna()]
print('number to geocode after matching:', len(df_lat))

# compare matches across provinces

# for each province want
# - number of rows
# - percent perfect matches
# - percent matches 
# - remainder with no geocodes

provinces = ['AB', 'BC', 'MB', 'NB', 'NT', 'NS', 'ON', 'PE', 'QC', 'SK']

for province in provinces:
    df2 = df[df['province'] == province]
    no_matches = df2['ratio'][df2['ratio'] > 80].count()
    perfect_matches = df2['ratio'][df2['ratio'] > 99].count()
    percent_matches = int(100 * no_matches / len(df2))
    percent_perfect_matches = int(100 * perfect_matches / len(df2))
#     print('Matches for ', province, ': ', int(percent_matches), '%', sep="")

    # get the number left to geocode - either an na in either 'longitude' or 'x'

    print(province, ":", sep="")
    print('   rows: ', len(df2), sep="")
    print('   perfect matches: ', percent_perfect_matches, '%', sep="")
    print('   matches above 0.8: ', percent_matches, '%', sep="")
#     print('remaining to geocode: ', len(df2), sep="")
